renamed settings handler
